Changed dataSize parameter to size_t

+++ modules/uas_message/image_tagger_message.hpp
@@ -29,7 +29,7 @@ class ImageTaggerMessage: public UASMessage {
          * \param [in] imageData[] a byte array holding the data of the current image
          * \param [in] dataSize an int indicating the number of elements in the array
          */
-        ImageTaggerMessage(unsigned char sequenceNumber, unsigned char imageData[], int dataSize);
+        ImageTaggerMessage(unsigned char sequenceNumber, unsigned char imageData[], size_t dataSize);
 
         /*!
          * ~ImageTaggerMessage deconstructor
